requirements.txt
aiohttp
aiosqlite
aiofiles
discord.py
cachetools
easy-pil
Pillow
dotenv
